build: move tests out of source directory and exclude from wheel distribution
Fixes #313
