+++ MANIFEST.in
@@ -4,3 +4,4 @@ recursive-include treebeard/static *
 recursive-include treebeard/templates *
 recursive-include treebeard/locale *
 recursive-include docs Makefile README.md make.bat *.py *.png *.rst
+recursive-include tests *.py
