+++ pytest.ini
@@ -1,4 +1,4 @@
 [pytest]
-DJANGO_SETTINGS_MODULE = treebeard.tests.settings
+DJANGO_SETTINGS_MODULE = tests.settings
 python_files = tests.py test_*.py *_tests.py
 django_find_project = false
