+++ tests/admin.py
@@ -4,7 +4,7 @@ from django.contrib import admin
 from treebeard.admin import admin_factory
 from treebeard.forms import movenodeform_factory
 
-from treebeard.tests.models import BASE_MODELS, UNICODE_MODELS, DEP_MODELS
+from tests.models import BASE_MODELS, UNICODE_MODELS, DEP_MODELS
 
 
 def register(admin_site, model):
